refactor: continue migrating over multiple buses
