+++ src/bus/mod.rs
@@ -23,7 +23,8 @@ use crate::ServiceId;
 pub use client::Client;
 #[cfg(feature = "shell")]
 pub use request::OfferStatusSelector;
-pub use request::{Failure, FailureCode, Request};
+pub use request::{Request};
+pub use rpc::{Failure, FailureCode};
 
 use microservices::esb::BusId;
 use strict_encoding::{StrictDecode, StrictEncode};
