+++ src/swapd/swap_state.rs
@@ -2,7 +2,8 @@ use farcaster_core::role::{SwapRole, TradeRole};
 use strict_encoding::{StrictDecode, StrictEncode};
 
 use crate::bus::msg::Commit;
-use crate::bus::request::{Outcome, Params};
+use crate::bus::request::{Outcome};
+use crate::bus::ctl::Params;
 
 #[derive(Display, Debug, Clone, StrictEncode, StrictDecode)]
 pub enum AliceState {
